requirements.txt
asgiref
certifi
charset-normalizer
# coreapi
# coreschema
Django
django-cors-headers
djangorestframework
gunicorn
idna
itypes
Jinja2
joblib
MarkupSafe
numpy
packaging
pandas
patsy
pyparsing
python-dateutil
python-decouple
pytz
PyYAML
requests
scikit-learn
scipy
six
sqlparse
statsmodels
threadpoolctl
uritemplate
urllib3
whitenoise
